matrix array changes
The separate characters are initialized once, and can be accessed by its
index. The symbols are enumerated so they refer to the corresponding
matrix so making phrases is easier. To-do: loop through messages of
different length properly. sizeof not working because it is an array of
pointers. need another solution