Animation Style
Added and changed bools and switch statements to it is easier to switch
between different animation styles. Added a frown to smile sequence...
It looks awful... need to improve the image. To-do: cycle through
animations a limited or unlimited amount of times, stop on a certain
frame or end on a different symbol altogether.