Cycles and Ending Cycles
You can now cycle through sequences a certain amount of times. The
beginning SPACE can now be removed(but I might just add a SPACE at the
end now...) since the pattern increments in the 'right' location now.
The next best thing to finding array size is just putting it in an
array, but now it calculated it by itself if you add it. numCycles = -1
for infinite cycles. The pattern increments too far so it tries to
access a value outside the array so I ghetto fixed it by reducing it by
one. In hindsight/fix later, pattern++ should be before the logic so it
gets changed after. Code for a 1 row buffer between sliding text if
needed. (Mainly for vert. scrolling so the chars are separated. As a
side note, 'W' looks ridiculous since it is aligned to the very left.)